Demo_code/requirements.txt
openai>=1.0.0
anthropic>=0.25.0
google-genai>=0.2.0
python-dotenv>=1.0.0
click>=8.0.0 